tests + megafix

+++ src/main/java/com/parachute/booking/flight/FlightServiceRemove.java
@@ -1,5 +1,6 @@
 package com.parachute.booking.flight;
 
+import com.parachute.booking.exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,6 +13,11 @@ public class FlightServiceRemove {
     private final FlightRepository flightRepository;
 
     public void removeFlightById(Long id) {
-        flightRepository.deleteById(id);
+
+        if (flightRepository.findById(id).isPresent()) {
+            flightRepository.deleteById(id);
+        } else {
+            throw new NotFoundException("Not found flight to delete with ID: " + id);
+        }
     }
 }
