+++ src/main/java/com/parachute/booking/pilot/PilotServiceRemove.java
@@ -1,5 +1,6 @@
 package com.parachute.booking.pilot;
 
+import com.parachute.booking.exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,6 +13,12 @@ public class PilotServiceRemove {
     private final PilotRepository pilotRepository;
 
     public void removePilotById(Long id) {
-        pilotRepository.deleteById(id);
+
+        if (pilotRepository.findById(id).isPresent()) {
+            pilotRepository.deleteById(id);
+        } else {
+            throw new NotFoundException("Not found pilot to delete with ID: " + id);
+        }
     }
+
 }
