+++ src/main/java/com/parachute/booking/plane/PlaneServiceRemove.java
@@ -1,5 +1,6 @@
 package com.parachute.booking.plane;
 
+import com.parachute.booking.exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,7 +13,13 @@ public class PlaneServiceRemove {
     private final PlaneRepository planeRepository;
 
     public void removePlaneById(Long id) {
-        planeRepository.deleteById(id);
+
+        if (planeRepository.findById(id).isPresent()) {
+
+            planeRepository.deleteById(id);
+        } else {
+            throw new NotFoundException("Not found plane to delete with ID: " + id);
+        }
     }
 
 }
